requirements.txt
streamlit>=1.20
google-genai

pandas
matplotlib
#tensorflow==2.16.2   # only if you must run TF in the space (can be heavy)
huggingface-hub
#opencv-python
python-multipart
